requirements.txt
requests
Pillow
numpy
opencv-python